# openapi.yaml
openapi: 3.0.3
info:
  title: 'Gym API'
  version: 1.0.0
  description: 'ระบบข้อมูลฟิตเนส (Fitness Management System) - gyms, trainers, members, memberships'
servers:
  -
    url: /api/v1
    description: 'Current server'
...
paths:
...
  /trainers:
    get:
      tags:
        - trainers
      summary: ดึงเทรนเนอร์ทั้งหมด
      description: คืนรายการเทรนเนอร์ทั้งหมด
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: สำเร็จ
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Trainer'
        '500':
          $ref: '#/components/responses/ServerError'
    post:
      tags:
        - trainers
      summary: เพิ่มเทรนเนอร์ใหม่
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerInput'
            example:
              name: Somchai
              specialty: 'Weight training'
              gym_id: 1
      responses:
        '201':
          description: เพิ่มเทรนเนอร์สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: สำเร็จ
                  data:
                    $ref: '#/components/schemas/Trainer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: ไม่พบข้อมูลที่อ้างอิง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'ไม่พบข้อมูลฟิสเนตตาม gym_id ที่ระบุ'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/ServerError'
  '/trainers/{id}':
    get:
      tags:
        - trainers
      summary: 'ดึงเทรนเนอร์ตาม id'
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: สำเร็จ
                  data:
                    $ref: '#/components/schemas/Trainer'
        '404':
          description: ไม่พบข้อมูล
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: ไม่พบข้อมูลเทรนเนอร์
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - trainers
      summary: 'แก้ไขเทรนเนอร์ตาม id'
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrainerInput'
            example:
              name: Somchai
              specialty: 'Weight training'
              gym_id: 1
      responses:
        '200':
          description: แก้ไขเทรนเนอร์สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: สำเร็จ
                  data:
                    $ref: '#/components/schemas/Trainer'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: 'ไม่พบข้อมูล หรือข้อมูลที่อ้างอิง'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: ไม่พบข้อมูลเทรนเนอร์
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '500':
          $ref: '#/components/responses/ServerError'
    delete:
      tags:
        - trainers
      summary: 'ลบเทรนเนอร์ตาม id'
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
          example: 1
      responses:
        '200':
          description: ลบเทรนเนอร์สำเร็จ
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: ลบเทรนเนอร์สำเร็จ
        '404':
          description: ไม่พบข้อมูล
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: ไม่พบข้อมูลเทรนเนอร์
        '409':
          description: 'ลบไม่ได้เพราะมีข้อมูลอื่นอ้างอิงอยู่ (Foreign Key)'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: ไม่สามารถลบเทรนเนอร์นี้ได้
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    BadRequest:
      description: ข้อมูลไม่ครบถ้วนหรือไม่ถูกต้อง
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: ข้อมูลไม่ครบถ้วน
...
    UnsupportedMediaType:
      description: 'Content-Type ต้องเป็น application/json'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorObject'
          example:
            error:
              code: UNSUPPORTED_MEDIA_TYPE
              message: 'กรุณาส่งข้อมูลในรูปแบบ application/json'
    ServerError:
      description: ข้อผิดพลาดภายในระบบ
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorObject'
          example:
            error:
              code: INTERNAL_SERVER_ERROR
              message: เกิดข้อผิดพลาดที่ไม่คาดคิดภายในระบบ
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
    ErrorObject:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    TrainerInput:
      type: object
      required:
        - name
        - gym_id
      properties:
        name:
          type: string
        specialty:
          type: string
          nullable: true
        gym_id:
          type: integer
          example: 1
    Trainer:
      allOf:
        -
          type: object
          properties:
            id:
              type: integer
              example: 1
        -
          $ref: '#/components/schemas/TrainerInput'
...